Finished refactoring of edit form
